fix: Migrate In50Hours page to Next.js
