+++ tailwind.config.ts
@@ -8,9 +8,6 @@ const config: Config = {
   ],
   theme: {
     extend: {
-      fontFamily: {
-        poppins: ["var(--font-poppins)", "sans-serif"],
-      },
       colors: {
         "mulearn-blue": "#5570F1",
       },
